move code to emerging
